divide back and front
