+++ src/App.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import {BrowserRouter as Router, Route, Routes} from 'react-router-dom';
-import {Header} from './components/header/Header';
-import {CurrencyConvert} from './pages/currecyConvert/CurrencyConvert';
-import {NotFound} from './pages/NotFound';
-import {Rates} from './pages/rates/Rates';
-import {Registration} from './pages/registry/Registration';
-import { Autorisation } from './pages/autorisation/Autorisation';
-import {Footer} from './components/footer/Footer';
+import {Header} from './front/components/header/Header';
+import {CurrencyConvert} from './front/pages/currecyConvert/CurrencyConvert';
+import {NotFound} from './front/pages/NotFound';
+import {Rates} from './front/pages/rates/Rates';
+import {Registration} from './front/pages/registry/Registration';
+import {Autorisation} from './front/pages/autorisation/Autorisation';
+import {Footer} from './front/components/footer/Footer';
 import './App.css';
