+++ src/front/api/CurrencyApi.ts
