+++ src/front/pages/rates/Rates.css
