+++ src/front/pages/registry/Registration.css
